same ui interaction for pined messages
