+++ app/src/main/java/chat/rocket/android/chatroom/presentation/ChatRoomNavigator.kt
@@ -14,6 +14,12 @@ class ChatRoomNavigator(internal val activity: ChatRoomActivity) {
         }
     }
 
+    fun toPinnedMessageList(chatRoomId: String, chatRoomType: String) {
+        activity.addFragmentBackStack("PinnedMessages", R.id.fragment_container){
+            chat.rocket.android.pinnedmessages.ui.newInstance(chatRoomId,chatRoomType)
+        }
+    }
+
     fun toNewServer() {
         activity.startActivity(activity.changeServerIntent())
         activity.finish()
