--- app/src/main/java/chat/rocket/android/chatroom/ui/PinnedMessagesActivity.kt
@@ -1,66 +1,0 @@
-package chat.rocket.android.chatroom.ui
-
-import android.os.Bundle
-import android.support.v4.app.Fragment
-import android.support.v7.app.AppCompatActivity
-import chat.rocket.android.R
-import chat.rocket.android.util.extensions.addFragment
-import chat.rocket.android.util.extensions.textContent
-import dagger.android.AndroidInjection
-import dagger.android.AndroidInjector
-import dagger.android.DispatchingAndroidInjector
-import dagger.android.support.HasSupportFragmentInjector
-import kotlinx.android.synthetic.main.app_bar_chat_room.*
-import javax.inject.Inject
-
-private const val INTENT_CHAT_ROOM_ID = "chat_room_id"
-private const val INTENT_CHAT_ROOM_NAME = "chat_room_name"
-private const val INTENT_CHAT_ROOM_TYPE = "chat_room_type"
-
-class PinnedMessagesActivity : AppCompatActivity(), HasSupportFragmentInjector {
-    @Inject lateinit var fragmentDispatchingAndroidInjector: DispatchingAndroidInjector<Fragment>
-    private lateinit var chatRoomId: String
-    private lateinit var chatRoomName: String
-    private lateinit var chatRoomType: String
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        AndroidInjection.inject(this)
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_pinned_messages)
-
-        chatRoomId = intent.getStringExtra(INTENT_CHAT_ROOM_ID)
-        requireNotNull(chatRoomId) { "no chat_room_id provided in Intent extras" }
-
-        chatRoomName = intent.getStringExtra(INTENT_CHAT_ROOM_NAME)
-        requireNotNull(chatRoomName) { "no chat_room_name provided in Intent extras" }
-
-        chatRoomType = intent.getStringExtra(INTENT_CHAT_ROOM_TYPE)
-        requireNotNull(chatRoomType) { "no chat_room_type provided in Intent extras" }
-
-        setupToolbar()
-
-        addFragment("PinnedMessagesFragment", R.id.fragment_container) {
-            newPinnedMessagesFragment(chatRoomId, chatRoomName, chatRoomType)
-        }
-    }
-
-    override fun onBackPressed() = finishActivity()
-
-    override fun supportFragmentInjector(): AndroidInjector<Fragment> {
-        return fragmentDispatchingAndroidInjector
-    }
-
-    private fun setupToolbar() {
-        setSupportActionBar(toolbar)
-        supportActionBar?.setDisplayShowTitleEnabled(false)
-        text_room_name.textContent = getString(R.string.title_pinned_messages)
-        toolbar.setNavigationOnClickListener {
-            finishActivity()
-        }
-    }
-
-    private fun finishActivity() {
-        super.onBackPressed()
-        overridePendingTransition(R.anim.close_enter, R.anim.close_exit)
-    }
-}
