+++ app/src/main/java/chat/rocket/android/dagger/module/ActivityBuilder.kt
@@ -12,7 +12,6 @@ import chat.rocket.android.chatroom.di.ChatRoomFragmentProvider
 import chat.rocket.android.chatroom.di.ChatRoomModule
 import chat.rocket.android.chatroom.di.PinnedMessagesFragmentProvider
 import chat.rocket.android.chatroom.ui.ChatRoomActivity
-import chat.rocket.android.chatroom.ui.PinnedMessagesActivity
 import chat.rocket.android.chatrooms.di.ChatRoomsFragmentProvider
 import chat.rocket.android.dagger.scope.PerActivity
 import chat.rocket.android.main.di.MainModule
@@ -50,12 +49,9 @@ abstract class ActivityBuilder {
     @PerActivity
     @ContributesAndroidInjector(modules = [ChatRoomModule::class,
         ChatRoomFragmentProvider::class,
-        MembersFragmentProvider::class])
+        MembersFragmentProvider::class,
+        PinnedMessagesFragmentProvider::class])
     abstract fun bindChatRoomActivity(): ChatRoomActivity
-
-    @PerActivity
-    @ContributesAndroidInjector(modules = [PinnedMessagesFragmentProvider::class])
-    abstract fun bindPinnedMessagesActivity(): PinnedMessagesActivity
 
     @PerActivity
     @ContributesAndroidInjector(modules = [PasswordFragmentProvider::class])
